compute_ctc_loss to utils/common.py
